+++ cmd/local/threading/thread_pool.go
@@ -26,45 +26,53 @@ import (
 type ThreadPool struct {
 	wg               sync.WaitGroup
 	numberThreads    int
-	jobs             chan func() error
+	jobs             chan Job
 	pendingJobs      int
 	totalJobs        int
 	loggingFrequency int
 	output           bool
+	outputProgress   bool
 	mut              sync.Mutex
 }
 
-func NewThreadPool(numberThreads int, loggingFrequency int, output bool) (*ThreadPool, error) {
+func NewThreadPool(numberThreads int, loggingFrequency int, output bool, outputProgress bool) (*ThreadPool, error) {
 	if numberThreads == 0 {
 		return &ThreadPool{}, errors.New("invalid number of threads at 0")
 	}
 
 	//by default support 4 million jobs
-	jobs := make(chan func() error, 4000000)
+	jobs := make(chan Job, 4000000)
 	return &ThreadPool{
 		numberThreads:    numberThreads,
 		jobs:             jobs,
 		loggingFrequency: loggingFrequency,
 		output:           output,
+		outputProgress:   outputProgress,
 	}, nil
 }
 
-func NewThreadPoolWithJobQueue(numberThreads, jobQueueSize int, loggingFrequency int, output bool) (*ThreadPool, error) {
+func NewThreadPoolWithJobQueue(numberThreads, jobQueueSize int, loggingFrequency int, output bool, outputProgress bool) (*ThreadPool, error) {
 	if numberThreads == 0 {
 		return &ThreadPool{}, errors.New("invalid number of threads at 0")
 	}
-	jobs := make(chan func() error, jobQueueSize)
+	jobs := make(chan Job, jobQueueSize)
 
 	return &ThreadPool{
 		numberThreads:    numberThreads,
 		jobs:             jobs,
 		loggingFrequency: loggingFrequency,
 		output:           output,
+		outputProgress:   outputProgress,
 	}, nil
 }
 
+type Job struct {
+	Name    string
+	Process func() error
+}
+
 // AddJob adds a job to the thread pool. It increases the wait group counter and sends the job to the jobs channel.
-func (t *ThreadPool) AddJob(job func() error) {
+func (t *ThreadPool) AddJob(job Job) {
 	t.mut.Lock()
 	t.pendingJobs++
 	t.totalJobs++
@@ -76,22 +84,34 @@ func (t *ThreadPool) AddJob(job func() error) {
 // worker listens for jobs on the jobs channel and executes them. Each job runs on its own goroutine.
 func (t *ThreadPool) worker() {
 	for job := range t.jobs {
-		err := job()
+		msg := fmt.Sprintf("JOB START - %v", job.Name)
+		if t.output {
+			fmt.Println(msg)
+		}
+		simplelog.Info(msg)
+		err := job.Process()
 		if err != nil {
+			msg := fmt.Sprintf("JOB FAILED - %v - %v", job.Name, err)
 			if t.output {
-				fmt.Print("x")
+				fmt.Println(msg)
 			}
-			simplelog.Errorf("Failed to execute job: %v", err)
+			simplelog.Error(msg)
 		} else {
+			msg := fmt.Sprintf("JOB COMPLETE - %v", job.Name)
 			if t.output {
-				fmt.Print(".")
+				fmt.Println(msg)
 			}
+			simplelog.Info(msg)
 		}
 		t.mut.Lock()
 		t.pendingJobs--
 		jobsCompleted := t.totalJobs - t.pendingJobs
 		if jobsCompleted%t.loggingFrequency == 0 {
-			simplelog.Infof("%v/%v tasks completed", jobsCompleted, t.totalJobs)
+			msg := fmt.Sprintf("JOB PROGRESS - %.2f%% COMPLETED", (float64(t.totalJobs-t.pendingJobs)*100.0)/float64(t.totalJobs))
+			if t.outputProgress {
+				fmt.Println(msg)
+			}
+			simplelog.Info(msg)
 		}
 		t.mut.Unlock()
 		t.wg.Done()
@@ -114,7 +134,11 @@ func (t *ThreadPool) ProcessAndWait() error {
 	t.wg.Wait()
 	close(t.jobs)
 	t.mut.Lock()
-	simplelog.Infof("%v/%v tasks completed", t.totalJobs, t.totalJobs)
+	msg := fmt.Sprintf("JOB PROGRESS - %.2f%% COMPLETED", (float64(t.totalJobs-t.pendingJobs)*100.0)/float64(t.totalJobs))
+	if t.outputProgress {
+		fmt.Println(msg)
+	}
+	simplelog.Info(msg)
 	t.totalJobs = 0
 	t.mut.Unlock()
 	return nil